8995 oxinstaller return value is always 0

+++ etc/oxfunctions.sh
@@ -136,6 +136,7 @@ ox_set_property() {
     test -z "$prop"     && die "ox_set_property: missing prop argument (arg 1)"
     test -z "$val"      && die "ox_set_property: missing val argument (arg 2)"
     test -z "$propfile" && die "ox_set_property: missing propfile argument (arg 3)"
+    test -e "$propfile" || die "ox_set_property: $propfile does not exist"
     local tmp=${propfile}.tmp$$
     rm -f $tmp
     # quote & in URLs to make sed happy
